+++ Sources/ZcashLightClientKit/Transaction/TransactionEncoder.swift
@@ -19,44 +19,58 @@ public enum TransactionEncoderError: Error {
 }
 
 protocol TransactionEncoder {
-    /// Creates a transaction, throwing an exception whenever things are missing. When the provided wallet implementation
-    /// doesn't throw an exception, we wrap the issue into a descriptive exception ourselves (rather than using
-    /// double-bangs for things).
+    /// Creates a proposal for transferring funds to the given recipient.
     ///
-    /// - Parameters:
-    /// - Parameter spendingKey: a `UnifiedSpendingKey` containing the spending key
-    /// - Parameter zatoshi: the amount to send in `Zatoshi`
-    /// - Parameter to: string containing the recipient address
-    /// - Parameter MemoBytes: string containing the memo (optional)
-    /// - Parameter accountIndex: index of the account that will be used to send the funds
-    /// - Throws:
-    ///     - `walletTransEncoderCreateTransactionMissingSaplingParams` if the sapling parameters aren't downloaded.
-    ///     - Some `ZcashError.rust*` if the creation of transaction fails.
-    func createTransaction(
-        spendingKey: UnifiedSpendingKey,
-        zatoshi: Zatoshi,
-        to address: String,
-        memoBytes: MemoBytes?,
-        from accountIndex: Int
-    ) async throws -> ZcashTransaction.Overview
-    
-    /// Creates a transaction that will attempt to shield transparent funds that are present on the blocks cache .throwing
-    /// an exception whenever things are missing. When the provided wallet implementation doesn't throw an exception,
-    /// we wrap the issue into a descriptive exception ourselves (rather than using double-bangs for things).
+    /// - Parameter accountIndex: the account from which to transfer funds.
+    /// - Parameter recipient: string containing the recipient's address.
+    /// - Parameter amount: the amount to send in Zatoshi.
+    /// - Parameter memoBytes: an optional memo to include as part of the proposal's transactions. Use `nil` when sending to transparent receivers otherwise the function will throw an error.
     ///
-    /// - Parameters:
-    /// - Parameter spendingKey: `UnifiedSpendingKey` to spend the UTXOs
-    /// - Parameter memoBytes: containing the memo (optional)
-    /// - Parameter accountIndex: index of the account that will be used to send the funds
-    /// - Throws:
-    ///     - `walletTransEncoderShieldFundsMissingSaplingParams` if the sapling parameters aren't downloaded.
-    ///     - Some `ZcashError.rust*` if the creation of transaction fails.
-    func createShieldingTransaction(
-        spendingKey: UnifiedSpendingKey,
+    /// If `prepare()` hasn't already been called since creation of the synchronizer instance or since the last wipe then this method throws
+    /// `SynchronizerErrors.notPrepared`.
+    func proposeTransfer(
+        accountIndex: Int,
+        recipient: String,
+        amount: Zatoshi,
+        memoBytes: MemoBytes?
+    ) async throws -> Proposal
+
+    /// Creates a proposal for shielding any transparent funds received by the given account.
+    ///
+    /// - Parameter accountIndex: the account for which to shield funds.
+    /// - Parameter shieldingThreshold: the minimum transparent balance required before a proposal will be created.
+    /// - Parameter memoBytes: an optional memo to include as part of the proposal's transactions.
+    /// - Parameter transparentReceiver: a specific transparent receiver within the account
+    ///             that should be the source of transparent funds. Default is `nil` which
+    ///             will select whichever of the account's transparent receivers has funds
+    ///             to shield.
+    ///
+    /// Returns the proposal, or `nil` if the transparent balance that would be shielded
+    /// is zero or below `shieldingThreshold`.
+    ///
+    /// If `prepare()` hasn't already been called since creation of the synchronizer instance or since the last wipe then this method throws
+    /// `SynchronizerErrors.notPrepared`.
+    func proposeShielding(
+        accountIndex: Int,
         shieldingThreshold: Zatoshi,
         memoBytes: MemoBytes?,
-        from accountIndex: Int
-    ) async throws -> ZcashTransaction.Overview
+        transparentReceiver: String?
+    ) async throws -> Proposal?
+
+    /// Creates the transactions in the given proposal.
+    ///
+    /// - Parameter proposal: the proposal for which to create transactions.
+    /// - Parameter spendingKey: the `UnifiedSpendingKey` associated with the account for which the proposal was created.
+    /// - Throws:
+    ///     - `walletTransEncoderCreateTransactionMissingSaplingParams` if the sapling parameters aren't downloaded.
+    ///     - Some `ZcashError.rust*` if the creation of transaction(s) fails.
+    ///
+    /// If `prepare()` hasn't already been called since creation of the synchronizer instance or since the last wipe then this method throws
+    /// `SynchronizerErrors.notPrepared`.
+    func createProposedTransactions(
+        proposal: Proposal,
+        spendingKey: UnifiedSpendingKey
+    ) async throws -> [ZcashTransaction.Overview]
 
     /// submits a transaction to the Zcash peer-to-peer network.
     /// - Parameter transaction: a transaction overview